+++ web-frontend/modules/database/components/sidebar/Sidebar.vue
@@ -35,6 +35,14 @@
           :table="table"
         ></SidebarItem>
       </ul>
+      <ul v-if="pendingJobs.length" class="tree__subs">
+        <SidebarItemPendingJob
+          v-for="job in pendingJobs"
+          :key="job.id"
+          :job="job"
+        >
+        </SidebarItemPendingJob>
+      </ul>
       <a class="tree__sub-add" @click="$refs.importFileModal.show()">
         <i class="fas fa-plus"></i>
         {{ $t('sidebar.createTable') }}
@@ -47,6 +55,7 @@
 <script>
 import { notifyIf } from '@baserow/modules/core/utils/error'
 import SidebarItem from '@baserow/modules/database/components/sidebar/SidebarItem'
+import SidebarItemPendingJob from '@baserow/modules/database/components/sidebar/SidebarItemPendingJob'
 import ImportFileModal from '@baserow/modules/database/components/table/ImportFileModal'
 import SidebarApplication from '@baserow/modules/core/components/sidebar/SidebarApplication'
 
@@ -55,6 +64,7 @@ export default {
   components: {
     SidebarApplication,
     SidebarItem,
+    SidebarItemPendingJob,
     ImportFileModal,
   },
   props: {
@@ -72,6 +82,13 @@ export default {
       return this.application.tables
         .map((table) => table)
         .sort((a, b) => a.order - b.order)
+    },
+    pendingJobs() {
+      return this.$store.getters['job/getAll'].filter((job) =>
+        this.$registry
+          .get('job', job.type)
+          .isJobPartOfApplication(job, this.application)
+      )
     },
   },
   methods: {